Now working correctly
The two problems that caused it not to work are:
1. The wrong data size was being passed to insertionSort
2. The data was not being cast correctly in the callback comparsion function

With respect to point 1. Two allocations were done, one as an array of
pointers and then another allocation initializing each pointer in that
array. This means that we essentially pass an array of character pointers
to insertion sort where the size of each element in the array is sizeof(char *).

For point 2. We pass a pointer to the data type stored in the array to be
compared, this is of type char **. So we first need to cast the void * type
to char ** and then perform a dereference on char ** to get the original
string to compare.

With these two fixes, the program sorts correctly.